+++ cmd/data-node/node/node.go
@@ -98,12 +98,13 @@ type NodeCommand struct {
 	chainInfoStore        *storage.ChainInfo
 	transferStore         *storage.Transfers
 
-	sqlStore        *sqlstore.SqlStore
-	assetStoreSql   *sqlstore.Assets
-	blockStoreSql   *sqlstore.Blocks
-	accountStoreSql *sqlstore.Accounts
-	ledgerSql       *sqlstore.Ledger
-	partyStoreSql   *sqlstore.Parties
+	sqlStore        *sqlstore.SQLStore
+	assetStoreSQL   *sqlstore.Assets
+	blockStoreSQL   *sqlstore.Blocks
+	accountStoreSQL *sqlstore.Accounts
+	balanceStoreSQL *sqlstore.Balances
+	ledgerSQL       *sqlstore.Ledger
+	partyStoreSQL   *sqlstore.Parties
 
 	vegaCoreServiceClient vegaprotoapi.CoreServiceClient
 
@@ -133,9 +134,9 @@ type NodeCommand struct {
 	checkpointSub        *subscribers.CheckpointSub
 	transferSub          *subscribers.TransferSub
 
-	assetSubSql            *sqlsubscribers.Asset
-	timeSubSql             *sqlsubscribers.Time
-	transferResponseSubSql *sqlsubscribers.TransferResponse
+	assetSubSQL            *sqlsubscribers.Asset
+	timeSubSQL             *sqlsubscribers.Time
+	transferResponseSubSQL *sqlsubscribers.TransferResponse
 
 	candleService     *candles.Svc
 	tradeService      *trades.Svc
@@ -234,6 +235,7 @@ func (l *NodeCommand) runNode(args []string) error {
 		l.rewardsSub,
 		l.stakingService,
 		l.checkpointSvc,
+		l.balanceStoreSQL,
 	)
 
 	// watch configs
@@ -258,7 +260,7 @@ func (l *NodeCommand) runNode(args []string) error {
 		return l.broker.Receive(ctx)
 	})
 
-	if l.conf.SqlStore.Enabled {
+	if l.conf.SQLStore.Enabled {
 		eg.Go(func() error {
 			return l.sqlBroker.Receive(ctx)
 		})