+++ entities/market.go
@@ -1,0 +1,10 @@
+package entities
+
+import (
+	"time"
+)
+
+type Market struct {
+	ID       []byte
+	VegaTime time.Time
+}