+++ sqlstore/accounts_test.go
@@ -5,6 +5,7 @@ import (
 
 	"code.vegaprotocol.io/data-node/entities"
 	"code.vegaprotocol.io/data-node/sqlstore"
+	"code.vegaprotocol.io/vega/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -55,4 +56,34 @@ func TestAccount(t *testing.T) {
 	fetchedAccount, err := accountStore.GetByID(account.ID)
 	assert.NoError(t, err)
 	assert.Equal(t, account, fetchedAccount)
+
+	// Add a second account, same asset - different party
+	party2 := addTestParty(t, partyStore, block)
+	account2 := addTestAccount(t, accountStore, party2, asset, block)
+
+	// Query by asset, should have 2 accounts
+	filter := entities.AccountFilter{Asset: asset}
+	accs, err := accountStore.Query(filter)
+	assert.NoError(t, err)
+	assert.Len(t, accs, 2)
+
+	// Query by asset + party should have only 1 account
+	filter = entities.AccountFilter{Asset: asset, Parties: []entities.Party{party2}}
+	accs, err = accountStore.Query(filter)
+	assert.NoError(t, err)
+	assert.Len(t, accs, 1)
+	assert.Equal(t, accs[0], account2)
+
+	// Query by asset + invalid type, should have 0 accounts
+	filter = entities.AccountFilter{Asset: asset, AccountTypes: []types.AccountType{100}}
+	accs, err = accountStore.Query(filter)
+	assert.NoError(t, err)
+	assert.Len(t, accs, 0)
+
+	// Query by asset + invalid market, should have 0 accounts
+	filter = entities.AccountFilter{Asset: asset, Markets: []entities.Market{{ID: []byte("Not A Market")}}}
+	accs, err = accountStore.Query(filter)
+	assert.NoError(t, err)
+	assert.Len(t, accs, 0)
+
 }