+++ sqlstore/utils.go
@@ -1,0 +1,15 @@
+package sqlstore
+
+import "strconv"
+
+// A handy little helper function for building queries. Appends 'value'
+// to the 'args' slice and returns a string '$N' referring to the index
+// of the value in args. For example:
+//
+// 	var args []interface{}
+//  query = "select * from foo where id=" + nextBindVar(&args, 100)
+//  db.Query(query, args...)
+func nextBindVar(args *[]interface{}, value interface{}) string {
+	*args = append(*args, value)
+	return "$" + strconv.Itoa(len(*args))
+}